--- Examples/Examples-Core.csproj
@@ -1,52 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Binding\Binding.csproj" />
-    <ProjectReference Include="..\Xaml\Xaml.csproj" />
-    <ProjectReference Include="..\ConsoleFramework\ConsoleFramework.csproj" />
-  </ItemGroup>
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <RootNamespace>Examples_Core.csproj</RootNamespace>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <StartupObject>Examples.Program</StartupObject>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Include="GridTest.xml" LogicalName="Examples.GridTest.xml" />
-    <EmbeddedResource Include="MainWindow.xml" LogicalName="Examples.MainWindow.xml" />
-    <EmbeddedResource Include="Layout.xml" LogicalName="Examples.Layout.xml" />
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Include="AsyncUiUpdate\main.xml" LogicalName="Examples.AsyncUiUpdate.main.xml" />
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Include="CheckBoxes\main.xml" LogicalName="Examples.CheckBoxes.main.xml" />
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Include="MainMenu\main.xml" LogicalName="Examples.MainMenu.main.xml" />
-    <EmbeddedResource Include="MainMenu\windows-host.xml" LogicalName="Examples.MainMenu.windows-host.xml" />
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Include="Commands\main.xml" LogicalName="Examples.Commands.main.xml" />
-    <EmbeddedResource Include="Commands\windows-host.xml" LogicalName="Examples.Commands.windows-host.xml" />
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Include="TreeView\main.xml" LogicalName="Examples.TreeView.main.xml" />
-    <EmbeddedResource Include="TreeView\windows-host.xml" LogicalName="Examples.TreeView.windows-host.xml" />
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Include="RadioButtons\main.xml" LogicalName="Examples.RadioButtons.main.xml" />
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Include="TabControl\main.xml" LogicalName="Examples.TabControl.main.xml" />
-  </ItemGroup>
-
-</Project>
