--- Examples/MainWindow.xml
@@ -1,31 +1,0 @@
-﻿<my:Window Name="window2" Title="Очень длинное название окна"
-        xmlns:x="http://consoleframework.org/xaml.xsd"
-        xmlns:my="clr-namespace:ConsoleFramework.Controls;assembly=ConsoleFramework"
-        xmlns:converters="clr-namespace:Binding.Converters;assembly=Binding"
-        xmlns:xaml="clr-namespace:ConsoleFramework.Xaml;assembly=ConsoleFramework">
-  <Window.Resources>
-    <string x:Key="1">Строка</string>
-    <converters:StringToIntegerConverter x:Key="2" x:Id="str2int"></converters:StringToIntegerConverter>
-  </Window.Resources>
-  <my:Panel>
-    <TextBox Text="{Binding Path=Str}" MaxLength="{Convert Converter={Ref str2int}, Value={Ref testStr}}"></TextBox>
-    <GroupBox Title="{Ref Ref=testStr}">
-      <ScrollViewer VerticalAlignment="Stretch">
-        <ScrollViewer.HorizontalScrollEnabled>false</ScrollViewer.HorizontalScrollEnabled>
-        <ListBox>
-          <ListBox.Resources>
-            <string x:Key="testItem" x:Id="testStr">5</string>
-          </ListBox.Resources>
-          <ListBox.Items>
-            <string>{Ref testStr}</string>
-            <string>Длинный элемент 3</string>
-            <string>Длинный элемент 4</string>
-            <string>Длинный элемент 5</string>
-            <string>Длинный элемент 6</string>
-            <string>Длинный элемент 700</string>
-          </ListBox.Items>
-        </ListBox>
-      </ScrollViewer>
-    </GroupBox>
-  </my:Panel>
-</my:Window>
